+++ src/dune_rules/check_rules.mli
@@ -1,10 +1,7 @@
 open Import
 
 val add_obj_dir :
-     Super_context.t
-  -> obj_dir:Path.Build.t Obj_dir.t
-  -> [ `Melange | `Ocaml ]
-  -> unit Memo.t
+  Super_context.t -> obj_dir:Path.Build.t Obj_dir.t -> Lib_mode.t -> unit Memo.t
 
 val add_files :
   Super_context.t -> dir:Path.Build.t -> Path.t list -> unit Memo.t