+++ test/blackbox-tests/test-cases/melange/add-rules-under-melange-emit-target.t
@@ -7,7 +7,7 @@ Test that the target directory exists
 
   $ cat > dune <<EOF
   > (melange.emit
-  >  (alias melange)
+  >  (alias mel)
   >  (emit_stdlib false)
   >  (target output))
   > EOF
@@ -18,13 +18,13 @@ Create the target dir
   $ cat > output/dune <<EOF
   > (rule
   >  (with-stdout-to index.txt (echo "hello")))
-  > (alias (name melange) (deps index.txt))
+  > (alias (name mel) (deps index.txt))
   > EOF
   $ cat > hello.ml <<EOF
   > let () = Js.log "hello"
   > EOF
 
-  $ dune build @melange
+  $ dune build @mel
   $ ls _build/default/output
   hello.js
   index.txt
@@ -34,13 +34,13 @@ Target promotion works
   $ dune clean
   $ cat > dune <<EOF
   > (melange.emit
-  >  (alias melange)
+  >  (alias mel)
   >  (target output)
   >  (emit_stdlib false)
   >  (promote (until-clean)))
   > EOF
 
-  $ dune build @melange
+  $ dune build @mel
   $ ls _build/default/output
   hello.js
   index.txt