+++ test/blackbox-tests/test-cases/melange/copy-files-include-subdirs.t
@@ -15,14 +15,14 @@ Example using melange.emit, copy_files and include_subdirs
   $ cat > src/dune <<EOF
   > (melange.emit
   >  (target app)
-  >  (alias melange))
+  >  (alias mel))
   > 
   > (subdir
   >  app
   >  (copy_files
   >   (files %{project_root}/assets/file.txt))
   >  (alias
-  >   (name melange)
+  >   (name mel)
   >   (deps file.txt)))
   > EOF
 
@@ -36,7 +36,7 @@ Example using melange.emit, copy_files and include_subdirs
   $ output_dir=_build/default/src/app
   $ src=$output_dir/src/main.js
   $ asset=$output_dir/file.txt
-  $ dune build @melange
+  $ dune build @mel
   $ dune build $asset
   $ node $src
   hello from file
@@ -46,7 +46,7 @@ Now add include_subdirs unqualified to show issue
 
   $ echo "(include_subdirs unqualified)" >> src/dune
 
-  $ dune build @melange
+  $ dune build @mel
   $ dune build $asset
   $ node $src
   hello from file