+++ test/blackbox-tests/test-cases/melange/ocaml-flags.t
@@ -10,7 +10,7 @@ Create dune file that uses melange.compile_flags
   $ cat > dune <<EOF
   > (melange.emit
   >  (target output)
-  >  (alias melange)
+  >  (alias mel)
   >  (modules main)
   >  (compile_flags -w -14-26))
   > EOF
@@ -24,7 +24,7 @@ The code in main contains unused var (warning 26) and illegal backlash (warning 
 
 Building does not fail, warnings are silenced
 
-  $ dune build @melange
+  $ dune build @mel
   $ node _build/default/output/main.js
   hello